Removed c++17 lexer and made new lexer compile with emscripten
